added a new file for the dress view and moved the view from main to the dress_view.dart, and used a grid view to display the items
